{{ cookiecutter.repo_name }}/requirements.txt
tensorflow
wandb
hydra-core